convert long and int to size_t if address space is addressed

+++ src/cato/rtlib/MemoryAbstraction.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <stdlib.h>
 
-MemoryAbstraction::MemoryAbstraction(long size, MPI_Datatype type, int dimensions)
+MemoryAbstraction::MemoryAbstraction(size_t size, MPI_Datatype type, int dimensions)
 {
     _base_ptr = nullptr;
     _size_bytes = size;
@@ -14,24 +14,24 @@ MemoryAbstraction::MemoryAbstraction(long size, MPI_Datatype type, int dimension
 
 MemoryAbstraction::~MemoryAbstraction() {}
 
-void MemoryAbstraction::store(void *base_ptr, void *value_ptr, std::vector<long> indices) {}
+void MemoryAbstraction::store(void *base_ptr, void *value_ptr, std::vector<size_t> indices) {}
 
-void MemoryAbstraction::load(void *base_ptr, void *dest_ptr, std::vector<long> indices) {}
+void MemoryAbstraction::load(void *base_ptr, void *dest_ptr, std::vector<size_t> indices) {}
 
 void MemoryAbstraction::sequential_store(void *base_ptr, void *value_ptr,
-                                         std::vector<long> indices)
+                                         std::vector<size_t> indices)
 {
 }
 
 void MemoryAbstraction::sequential_load(void *base_ptr, void *dest_ptr,
-                                        std::vector<long> indices)
+                                        std::vector<size_t> indices)
 {
 }
 
-void MemoryAbstraction::pointer_store(void *source_ptr, long dest_index) {}
+void MemoryAbstraction::pointer_store(void *source_ptr, size_t dest_index) {}
 
 void *MemoryAbstraction::get_base_ptr() { return _base_ptr; }
 
-long MemoryAbstraction::get_size_bytes() { return _size_bytes; }
+size_t MemoryAbstraction::get_size_bytes() { return _size_bytes; }
 
 MPI_Datatype MemoryAbstraction::get_type() { return _type; }
