+++ src/cato/rtlib/MemoryAbstraction.h
@@ -16,7 +16,7 @@ class MemoryAbstraction
   protected:
     void *_base_ptr;
 
-    long _size_bytes;
+    size_t _size_bytes;
 
     MPI_Datatype _type;
 
@@ -29,7 +29,7 @@ class MemoryAbstraction
      *
      * The constructor is responsible for the actual memory allocation
      **/
-    MemoryAbstraction(long size, MPI_Datatype type, int dimensions);
+    MemoryAbstraction(size_t size, MPI_Datatype type, int dimensions);
 
     /**
      * Destructor needs to free the allocated shared memory and
@@ -42,14 +42,14 @@ class MemoryAbstraction
      * This gets called from prallelized sections of the original
      * program.
      **/
-    virtual void store(void *base_ptr, void *value_ptr, std::vector<long> indices);
+    virtual void store(void *base_ptr, void *value_ptr, std::vector<size_t> indices);
 
     /**
      * A load from the shared memory object.
      * This gets called from prallelized sections of the original
      * program.
      **/
-    virtual void load(void *base_ptr, void *dest_ptr, std::vector<long> indices);
+    virtual void load(void *base_ptr, void *dest_ptr, std::vector<size_t> indices);
 
     /**
      * A store to the shared memory object.
@@ -58,7 +58,7 @@ class MemoryAbstraction
      * that it can be called by all MPI processes simultaniously and that
      * this is a synchronous operation between all processes.
      **/
-    virtual void sequential_store(void *base_ptr, void *value_ptr, std::vector<long> indices);
+    virtual void sequential_store(void *base_ptr, void *value_ptr, std::vector<size_t> indices);
 
     /**
      * A load from the shared memory object.
@@ -67,16 +67,16 @@ class MemoryAbstraction
      * that it can be called by all MPI processes simultaniously and that
      * this is a synchronous operation between all processes.
      **/
-    virtual void sequential_load(void *base_ptr, void *dest_ptr, std::vector<long> indices);
+    virtual void sequential_load(void *base_ptr, void *dest_ptr, std::vector<size_t> indices);
 
     /**
      * A pointer store to an MemoryAbstraction with pointer depth >= 2.
      **/
-    virtual void pointer_store(void *source_ptr, long dest_index);
+    virtual void pointer_store(void *source_ptr, size_t dest_index);
 
     virtual void *get_base_ptr();
 
-    virtual long get_size_bytes();
+    virtual size_t get_size_bytes();
 
     virtual MPI_Datatype get_type();
 };
